requirements.txt
python-telegram-bot
numpy
networkx
matplotlib
pulp